inital refactor of tools + routes
