+++ app.py
@@ -179,295 +179,9 @@ class SQLiteTracer(BaseCallbackHandler):
     def on_chain_end(self, outputs, **kwargs):
         self._write({"event":"chain_end","outputs":outputs})
 
-# =============================================================================
-# SUPER CAPABLE AGENT TOOLS
-# =============================================================================
 
-@tool
-def web_search(query: str, num_results: int = 5) -> str:
-    """Search the web using OpenAI's web search tool and return a concise, cited answer."""
-    try:
-        resp = client.responses.create(
-            model="gpt-4o",
-            input=(
-                f"Use web_search. Answer concisely, then list up to {num_results} sources with links."
-                f"\nQuery: {query}"
-            ),
-            tools=[{"type": "web_search"}],
-            tool_choice="auto",
-            max_output_tokens=600,
-        )
-        # Prefer SDK convenience field; fall back to manual parse
-        text = getattr(resp, "output_text", "") or ""
-        if not text and getattr(resp, "output", None):
-            parts = []
-            for item in resp.output:
-                if getattr(item, "type", "") == "output_text":
-                    for seg in getattr(item, "content", []) or []:
-                        if hasattr(seg, "text"):
-                            parts.append(seg.text)
-            text = "\n".join(parts).strip()
-        return text or "No results."
-    except Exception as e:
-        return f"Search unavailable: {e}"
-
-@tool
-def get_current_time(timezone: str = "UTC") -> str:
-    """Get current time in specified timezone (e.g., 'UTC', 'US/Eastern', 'Europe/London')."""
-    try:
-        # Try with pytz if available
-        try:
-            import pytz
-            if timezone.upper() == "UTC":
-                tz = pytz.UTC
-            else:
-                tz = pytz.timezone(timezone)
-            current_time = datetime.now(tz)
-            return f"Current time in {timezone}: {current_time.strftime('%Y-%m-%d %H:%M:%S %Z')}"
-        except ImportError:
-            # Fallback without pytz
-            current_time = datetime.utcnow()
-            return f"Current UTC time: {current_time.strftime('%Y-%m-%d %H:%M:%S UTC')} (pytz not available for timezone conversion)"
-    except Exception as e:
-        return f"Time error: {str(e)}"
-
-@tool
-def calculate_math(expression: str) -> str:
-    """Safely evaluate mathematical expressions. Supports +, -, *, /, **, sqrt, sin, cos, log, ln, etc.
-    Use 'ln(x)' for natural logarithm, 'log(x)' for base-10 logarithm.
-    
-    IMPORTANT: Provide ONLY the mathematical expression, NO comments or extra text."""
-    try:
-        # Debug logging
-        print(f"DEBUG: Original expression: '{expression}'")
-        
-        # Clean the input - remove comments and extra text
-        cleaned_expression = expression.strip()
-        
-        # Remove comments (anything after #)
-        if '#' in cleaned_expression:
-            cleaned_expression = cleaned_expression.split('#')[0].strip()
-        
-        # Remove quotes if present (both single and double quotes)
-        if (cleaned_expression.startswith('"') and cleaned_expression.endswith('"')) or \
-           (cleaned_expression.startswith("'") and cleaned_expression.endswith("'")):
-            cleaned_expression = cleaned_expression[1:-1]
-        
-        print(f"DEBUG: Cleaned expression: '{cleaned_expression}'")
-        
-        # Safe evaluation with math functions
-        allowed_names = {
-            k: v for k, v in math.__dict__.items() if not k.startswith("__")
-        }
-        allowed_names.update({
-            "abs": abs, "round": round, "min": min, "max": max,
-            "sum": sum, "len": len, "pow": pow,
-            "ln": math.log,  # Add natural logarithm as 'ln'
-        })
-        
-        # Compile and evaluate safely
-        code = compile(cleaned_expression, "<string>", "eval")
-        result = eval(code, {"__builtins__": {}}, allowed_names)
-        return f"Result: {result}"
-    except Exception as e:
-        return f"Math error: {str(e)}. Available functions: +, -, *, /, **, sqrt, sin, cos, log, ln, abs, round, min, max, etc. Make sure to provide only the mathematical expression without comments."
-
-@tool
-def analyze_data(data_str: str, operation: str = "describe") -> str:
-    """Analyze numerical data and return TEXT-BASED statistics only. Operations: 'describe', 'mean', 'median', 'std', 'min', 'max'. 
-    No visual charts or graphs - only text summaries and statistics."""
-    try:
-        # Parse numbers from string
-        numbers = [float(x.strip()) for x in re.findall(r'-?\d+\.?\d*', data_str)]
-        if not numbers:
-            return "No numerical data found in input"
-        
-        if operation == "describe":
-            return f"""Data Analysis:
-Count: {len(numbers)}
-Mean: {statistics.mean(numbers):.2f}
-Median: {statistics.median(numbers):.2f}
-Std Dev: {statistics.stdev(numbers) if len(numbers) > 1 else 0:.2f}
-Min: {min(numbers)}
-Max: {max(numbers)}
-Range: {max(numbers) - min(numbers)}"""
-        elif operation == "mean":
-            return f"Mean: {statistics.mean(numbers):.2f}"
-        elif operation == "median":
-            return f"Median: {statistics.median(numbers):.2f}"
-        elif operation == "std":
-            return f"Standard Deviation: {statistics.stdev(numbers) if len(numbers) > 1 else 0:.2f}"
-        elif operation == "min":
-            return f"Minimum: {min(numbers)}"
-        elif operation == "max":
-            return f"Maximum: {max(numbers)}"
-        else:
-            return f"Unknown operation: {operation}"
-    except Exception as e:
-        return f"Data analysis error: {str(e)}"
-
-@tool
-def read_file(file_path: str) -> str:
-    """Read and return the contents of a text file."""
-    try:
-        with open(file_path, 'r', encoding='utf-8') as f:
-            content = f.read()
-        return f"File content ({len(content)} characters):\n{content[:2000]}{'...' if len(content) > 2000 else ''}"
-    except Exception as e:
-        return f"Error reading file: {str(e)}"
-
-@tool
-def write_file(file_path: str, content: str) -> str:
-    """Write content to a text file."""
-    try:
-        with open(file_path, 'w', encoding='utf-8') as f:
-            f.write(content)
-        return f"Successfully wrote {len(content)} characters to {file_path}"
-    except Exception as e:
-        return f"Error writing file: {str(e)}"
-
-@tool
-def list_directory(path: str = ".") -> str:
-    """List files and directories in the specified path."""
-    try:
-        items = os.listdir(path)
-        dirs = [f"📁 {item}/" for item in items if os.path.isdir(os.path.join(path, item))]
-        files = [f"📄 {item}" for item in items if os.path.isfile(os.path.join(path, item))]
-        all_items = sorted(dirs) + sorted(files)
-        return f"Directory listing for '{path}':\n" + "\n".join(all_items[:50])
-    except Exception as e:
-        return f"Error listing directory: {str(e)}"
-
-@tool
-def execute_python_code(code: str) -> str:
-    """Execute Python code safely and return TEXT OUTPUT ONLY. 
-    
-    IMPORTANT INPUT FORMAT: Provide only the raw Python code, WITHOUT any markdown backticks or formatting.
-    
-    CORRECT: import math; print(math.sqrt(16))
-    WRONG: ```python\\nimport math\\nprint(math.sqrt(16))\\n```
-    
-    Use for calculations, data processing, and analysis. DO NOT use for plots/charts."""
-    try:
-        # Clean up the input - remove common formatting issues
-        cleaned_code = code.strip()
-        
-        # Remove markdown code block formatting if present
-        if cleaned_code.startswith('```'):
-            lines = cleaned_code.split('\n')
-            # Remove first line if it's ```python or similar
-            if lines[0].startswith('```'):
-                lines = lines[1:]
-            # Remove last line if it's ```
-            if lines and lines[-1].strip() == '```':
-                lines = lines[:-1]
-            cleaned_code = '\n'.join(lines)
-        
-        # Add encoding declaration for international characters
-        if '# -*- coding:' not in cleaned_code:
-            cleaned_code = "# -*- coding: utf-8 -*-\n" + cleaned_code
-        
-        # Create a temporary file
-        with tempfile.NamedTemporaryFile(mode='w', suffix='.py', delete=False, encoding='utf-8') as f:
-            f.write(cleaned_code)
-            temp_file = f.name
-        
-        # Execute with timeout
-        result = subprocess.run([
-            'python', temp_file
-        ], capture_output=True, text=True, timeout=30, encoding='utf-8')
-        
-        # Clean up
-        os.unlink(temp_file)
-        
-        output = result.stdout + result.stderr
-        
-        # Check for common formatting errors and provide helpful feedback
-        if "SyntaxError" in output and "```" in code:
-            return "Code execution error: You included markdown backticks (```) in your code. Please provide only raw Python code without any markdown formatting."
-        
-        return f"Code execution result:\n{output[:1000]}{'...' if len(output) > 1000 else ''}"
-    except subprocess.TimeoutExpired:
-        return "Code execution timed out (30s limit)"
-    except Exception as e:
-        return f"Code execution error: {str(e)}"
-
-@tool
-def get_weather(location: str) -> str:
-    """Get current weather information for a location using a free weather API."""
-    try:
-        # Using OpenWeatherMap free tier (you'd need to sign up for API key)
-        # For demo, returning mock data
-        import random
-        temperatures = list(range(-10, 35))
-        conditions = ["sunny", "cloudy", "rainy", "snowy", "partly cloudy", "overcast"]
-        
-        temp = random.choice(temperatures)
-        condition = random.choice(conditions)
-        humidity = random.randint(30, 90)
-        
-        return f"Weather in {location}:\nTemperature: {temp}°C\nCondition: {condition}\nHumidity: {humidity}%\n(Note: This is mock data for demo)"
-    except Exception as e:
-        return f"Weather error: {str(e)}"
-
-@tool
-def translate_text(text: str, target_language: str = "Spanish") -> str:
-    """Translate text to target language using a simple translation service."""
-    try:
-        # For demo purposes, providing basic translations
-        translations = {
-            "hello": {"Spanish": "hola", "French": "bonjour", "German": "hallo", "Italian": "ciao"},
-            "goodbye": {"Spanish": "adiós", "French": "au revoir", "German": "auf wiedersehen", "Italian": "ciao"},
-            "thank you": {"Spanish": "gracias", "French": "merci", "German": "danke", "Italian": "grazie"},
-        }
-        
-        text_lower = text.lower()
-        if text_lower in translations and target_language in translations[text_lower]:
-            return f"Translation to {target_language}: {translations[text_lower][target_language]}"
-        else:
-            return f"Translation service not available for '{text}' to {target_language}. Consider using Google Translate API."
-    except Exception as e:
-        return f"Translation error: {str(e)}"
-
-@tool
-def url_fetch(url: str) -> str:
-    """Fetch content from a URL and return a summary."""
-    try:
-        response = requests.get(url, timeout=10, headers={
-            'User-Agent': 'Mozilla/5.0 (compatible; AI Agent/1.0)'
-        })
-        response.raise_for_status()
-        content = response.text[:2000]  # Limit content
-        return f"URL content (first 2000 chars):\n{content}"
-    except Exception as e:
-        return f"URL fetch error: {str(e)}"
-
-@tool
-def json_processor(json_data: str, operation: str = "format") -> str:
-    """Process JSON data. Operations: 'format', 'validate', 'keys', 'values'."""
-    try:
-        data = json.loads(json_data)
-        
-        if operation == "format":
-            return json.dumps(data, indent=2)
-        elif operation == "validate":
-            return "Valid JSON format"
-        elif operation == "keys":
-            if isinstance(data, dict):
-                return f"JSON keys: {list(data.keys())}"
-            else:
-                return "Not a JSON object (no keys)"
-        elif operation == "values":
-            if isinstance(data, dict):
-                return f"JSON values: {list(data.values())}"
-            else:
-                return f"JSON data: {data}"
-        else:
-            return f"Unknown operation: {operation}"
-    except json.JSONDecodeError as e:
-        return f"Invalid JSON: {str(e)}"
-    except Exception as e:
-        return f"JSON processing error: {str(e)}"
+# Import agent tools from lib.tools.agent_tools
+from lib.tools.agent_tools import *
 
 # Note: get_conversation_history function removed as we now use messages array directly
 # instead of reconstructing history from traces
@@ -734,236 +448,16 @@ Provide a relevant, helpful response that addresses the user's question in the c
         conn.commit()
         conn.close()
 
-@app.route('/chat/new', methods=['POST'])
-def chat():
-    """
-    Handle chat requests with simplified format
-    Accepts: prompt (required) and session_id (optional)
-    Returns session_id and message_id for the assistant response
-    """
-    data = request.get_json()
-    
-    if not data or 'prompt' not in data:
-        return jsonify({"error": "Missing 'prompt' in request body"}), 400
-    
-    prompt = data['prompt']
-    session_id = data.get('session_id')  # Optional session_id for continuing conversations
-    
-    # Validate prompt
-    if not isinstance(prompt, str) or not prompt.strip():
-        return jsonify({"error": "'prompt' must be a non-empty string"}), 400
-    
-    conn = sqlite3.connect(DB_PATH)
-    cursor = conn.cursor()
-    
-    try:
-        # Handle existing vs new conversation
-        if session_id:
-            # Validate that the session exists
-            cursor.execute('SELECT session_id FROM sessions WHERE session_id = ?', (session_id,))
-            if not cursor.fetchone():
-                return jsonify({"error": "Session not found"}), 404
-        else:
-            # Create new conversation
-            session_id = str(uuid.uuid4())
-            cursor.execute('''
-                INSERT INTO sessions (session_id)
-                VALUES (?)
-            ''', (session_id,))
-        
-        # Store the new user message
-        user_message_id = str(uuid.uuid4())
-        cursor.execute('''
-            INSERT INTO messages (message_id, session_id, role, content, status)
-            VALUES (?, ?, 'user', ?, 'completed')
-        ''', (user_message_id, session_id, prompt.strip()))
-        
-        # Create a new assistant message for the response (initially pending)
-        assistant_message_id = str(uuid.uuid4())
-        cursor.execute('''
-            INSERT INTO messages (message_id, session_id, role, content, status)
-            VALUES (?, ?, 'assistant', '', 'pending')
-        ''', (assistant_message_id, session_id))
-        
-        conn.commit()
-        
-        # Build conversation context from database history (excluding current message and errors)
-        cursor.execute('''
-            SELECT role, content FROM messages 
-            WHERE session_id = ? AND message_id != ? AND status = 'completed'
-            ORDER BY created_at
-        ''', (session_id, user_message_id))
-        
-        history_messages = cursor.fetchall()
-        conversation_context = ""
-        
-        if history_messages:
-            context_parts = []
-            for role, content in history_messages:
-                if role == "user":
-                    context_parts.append(f"Human: {content}")
-                elif role == "assistant":
-                    context_parts.append(f"Assistant: {content}")
-                elif role == "system":
-                    context_parts.append(f"System: {content}")
-            
-            conversation_context = "Previous conversation:\n" + "\n".join(context_parts) + "\n\nCurrent request:\n"
-        
-        # Start agent execution in background thread
-        thread = threading.Thread(target=run_agent, args=(session_id, assistant_message_id, prompt.strip(), conversation_context))
-        thread.daemon = True
-        thread.start()
-        
-        return jsonify({
-            "session_id": session_id, 
-            "message_id": assistant_message_id
-        })
-        
-    except Exception as e:
-        conn.rollback()
-        return jsonify({"error": f"Database error: {str(e)}"}), 500
-    finally:
-        conn.close()
 
-@app.route('/sessions/<session_id>', methods=['GET'])
-def get_session(session_id):
-    """
-    Get all messages in a session
-    """
-    conn = sqlite3.connect(DB_PATH)
-    cursor = conn.cursor()
-    
-    # Get session info
-    cursor.execute('''
-        SELECT session_id, created_at, updated_at
-        FROM sessions WHERE session_id = ?
-    ''', (session_id,))
-    session_row = cursor.fetchone()
-    
-    if not session_row:
-        conn.close()
-        return jsonify({"error": "Session not found"}), 404
-    
-    # Get all messages for this session
-    cursor.execute('''
-        SELECT message_id, role, content, status, created_at, completed_at
-        FROM messages 
-        WHERE session_id = ?
-        ORDER BY created_at
-    ''', (session_id,))
-    message_rows = cursor.fetchall()
-    
-    conn.close()
-    
-    # Format response
-    session_info = {
-        "session_id": session_row[0],
-        "created_at": session_row[1],
-        "updated_at": session_row[2]
-    }
-    
-    messages = []
-    for message_id, role, content, status, created_at, completed_at in message_rows:
-        message = {
-            "message_id": message_id,
-            "role": role,
-            "content": content,
-            "status": status,
-            "created_at": created_at,
-            "completed_at": completed_at
-        }
-        messages.append(message)
-    
-    return jsonify({
-        "session": session_info,
-        "messages": messages
-    })
-
-@app.route('/messages/<message_id>/traces', methods=['GET'])
-def get_message_traces(message_id):
-    """
-    Get all traces for a specific message (assistant messages only)
-    """
-    conn = sqlite3.connect(DB_PATH)
-    cursor = conn.cursor()
-    
-    # Get message info
-    cursor.execute('''
-        SELECT message_id, session_id, role, content, status, created_at, completed_at
-        FROM messages WHERE message_id = ?
-    ''', (message_id,))
-    message_row = cursor.fetchone()
-    
-    if not message_row:
-        conn.close()
-        return jsonify({"error": "Message not found"}), 404
-    
-    # Only assistant messages have traces
-    if message_row[2] != 'assistant':  # role is at index 2
-        conn.close()
-        return jsonify({"error": "Only assistant messages have execution traces"}), 400
-    
-    # Get all traces for this message
-    cursor.execute('''
-        SELECT event, data, timestamp
-        FROM traces 
-        WHERE message_id = ?
-        ORDER BY timestamp
-    ''', (message_id,))
-    trace_rows = cursor.fetchall()
-    
-    conn.close()
-    
-    # Format response
-    message_info = {
-        "message_id": message_row[0],
-        "session_id": message_row[1],
-        "role": message_row[2],
-        "content": message_row[3],
-        "status": message_row[4],
-        "created_at": message_row[5],
-        "completed_at": message_row[6]
-    }
-    
-    traces = []
-    for event, data, timestamp in trace_rows:
-        try:
-            trace_data = json.loads(data)
-            traces.append(trace_data)
-        except json.JSONDecodeError:
-            traces.append({"event": event, "data": data, "timestamp": timestamp})
-    
-    return jsonify({
-        "message": message_info,
-        "traces": traces
-    })
-
-# Legacy endpoint for backward compatibility (deprecated)
-@app.route('/chat', methods=['GET'])
-def get_legacy_traces():
-    """
-    Legacy endpoint - redirects to new conversation-based API
-    """
-    session_id = request.args.get('session_id')
-    if session_id:
-        return jsonify({
-            "error": "This endpoint is deprecated. Use /sessions/{session_id} instead.",
-            "migration_note": "The new API uses session_id and message_id for better tracking"
-        }), 410  # Gone
-    else:
-        return jsonify({"error": "Missing 'session_id' parameter"}), 400
-
-@app.route('/', methods=['GET'])
-def health():
-    """Health check endpoint"""
-    return jsonify({"status": "healthy"})
+# Register routes from routes.api_routes
+from routes.api_routes import register_routes
 
 if __name__ == '__main__':
     # Initialize database
     init_db()
-    
+    # Register routes
+    register_routes(app)
     # Ensure OpenAI API key is set
     if not os.environ.get("OPENAI_API_KEY"):
         print("Warning: OPENAI_API_KEY environment variable not set")
-    
     app.run(debug=True, host='0.0.0.0', port=5000, use_reloader=True, reloader_type='stat')
