+++ routes/__init__.py
@@ -1,0 +1,1 @@
+# This file will contain all Flask route functions, to be registered with the app in app.py
